+++ CHANGELOG.rst
@@ -22,7 +22,7 @@ New features and notable changes:
 - Pygments required >= 2.13.0. (:issue:`799`)
 - Add a second theme for HTML report inspired by GitHub. (:issue:`793`)
 - Add :option:`--fail-under-decision` and :option:`--fail-under-function` which will error under a given minimum coverage. (:issue:`773`)
-
+- Add function coverage to data model. (:issue:`822`)
 
 Bug fixes and small improvements: